feat: Refactor line chart component to remove console log and improve chart rendering logic
feat: Update pie chart component HTML structure for better styling

feat: Revamp exam student component layout and add new columns for exam details

fix: Modify examen model to replace puntaje_letra with valor_examen

style: Clean up home component layout by removing unnecessary containers

style: Adjust margin in payment component for better UI alignment

fix: Update payment component to improve form handling and validation

feat: Enhance payment detail component with status icons for better UX

feat: Improve payment list component by adding status icons and cleaning up code

style: Refactor profile component layout for better readability

feat: Add calendar component to student detail view for enhanced functionality

feat: Implement calendar service for managing calendar tasks

test: Add unit tests for calendar component and service

feat: Create calendar model to manage calendar task data structure

fix: Update user service to include method for fetching teachers

+++ src/app/pages/students/student-list/student-list.component.ts
@@ -16,6 +16,7 @@ import {  RouterLink } from '@angular/router';
 import { ImagenPipe } from "../../../pipes/imagen.pipe";
 import { BackButtnComponent } from '../../../shared/backButtn/backButtn.component';
 import { PieChart2Component } from '../../../components/charts/pie-chart2/pie-chart2.component';
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-student-list',
   imports: [HeaderComponent,MenuFooterComponent,
@@ -111,14 +112,34 @@ export class StudentListComponent {
     }
 
     eliminarUser(user:Student){
-    this.studentService.deleteById(user.id).subscribe(
-      response =>{
-        this.getStudents();
-      },
-      error=>{
-        this.msm_error = 'No se pudo eliminar el curso, vuelva a intentar.'
-      }
-      );
-      this.ngOnInit();
+      Swal.fire({
+          title: "Esta Seguro?",
+          text: "Se perderán todos los datos!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Si, Borrar!"
+        }).then((result) => {
+          if (result.isConfirmed) {
+             this.studentService.deleteById(user.id).subscribe(
+              response =>{
+                Swal.fire({
+                  title: "Borrado!",
+                  text: "Registro Eliminado.",
+                  icon: "success"
+                });
+                
+                this.getStudents();
+              },
+              error=>{
+                this.msm_error = 'No se pudo eliminar el curso, vuelva a intentar.'
+              }
+            );
+            
+          }
+        });
+   
+      // this.getStudents();
   }
 }
